\name{npideal-package}
\alias{npideal-package}
\alias{npideal}
\docType{package}
\title{
\packageTitle{npideal}
}
\description{
\packageDescription{npideal}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{npideal}
\packageIndices{npideal}
}
\author{
\packageAuthor{npideal}

Maintainer: \packageMaintainer{npideal}
}
\references{
Tahk, Alexander. 2017. Nonparametric Ideal-Point Estimation and Inference.
}
\keyword{ package }
\examples{
}
